@@ -138,7 +138,7 @@ request.get('https://raw.githubusercontent.com/microsoft/react-native/master/pac
 
   let existingPkgJson = JSON.parse(fs.readFileSync(pkgJsonPath, 'utf8'));
 
-  const rnDependency = `${pkgJson.version} || https://github.com/microsoft/react-native/archive/v${pkgJson.version}.tar.gz`;
+  const rnDependency = `^${pkgJson.version.slice(0, pkgJson.version.indexOf('-'))} || ${pkgJson.version} || https://github.com/microsoft/react-native/archive/v${pkgJson.version}.tar.gz`;
 
   if (existingPkgJson.peerDependencies['react-native'] === rnDependency) {
     console.log(`Already at latest react-native version: ${pkgJson.version}.`);
