@@ -6,6 +6,7 @@
 #include "LocalBundleReader.h"
 #include "unicode.h"
 #include <winrt/Windows.Storage.h>
+#include <winrt/Windows.Storage.Streams.h>
 
 #if _MSC_VER <= 1913
 // VC 19 (2015-2017.6) cannot optimize co_await/cppwinrt usage
@@ -22,9 +23,20 @@ std::future<std::string> LocalBundleReader::LoadBundleAsync(const std::string& b
   co_await winrt::resume_background();
 
   auto file = co_await winrt::Windows::Storage::StorageFile::GetFileFromApplicationUriAsync(uri);
-  auto hdata = co_await winrt::Windows::Storage::FileIO::ReadTextAsync(file);
 
-  co_return facebook::react::unicode::utf16ToUtf8(std::wstring(hdata.data()));
+  // Read the buffer manually to avoid a Utf8 -> Utf16 -> Utf8 encoding roundtrip.
+  auto fileBuffer { co_await winrt::Windows::Storage::FileIO::ReadBufferAsync(file) };
+  auto dataReader{ winrt::Windows::Storage::Streams::DataReader::FromBuffer(fileBuffer) };
+
+  std::string script(fileBuffer.Length() + 1, '\0');
+
+  // Construct the array_view to slice into the first fileBuffer.Length bytes.
+  // DataReader.ReadBytes will read as many bytes as are present in the array_view.
+  // The backing string has fileBuffer.Length() + 1 bytes, without an explicit end it will read 1 byte to many and throw.
+  dataReader.ReadBytes(winrt::array_view<uint8_t> { reinterpret_cast<uint8_t*>(&script[0]), reinterpret_cast<uint8_t*>(&script[script.length() - 1]) });
+  dataReader.Close();
+
+  co_return script;
 }
 
 std::string LocalBundleReader::LoadBundle(const std::string& bundlePath)
