@@ -7,6 +7,7 @@
 
 #include <Folly/dynamic.h>
 #include <DevSettings.h>
+#include "XamlView.h"
 
 #include <functional>
 #include <string>
@@ -18,6 +19,7 @@ struct INativeUIManager;
 namespace react { namespace uwp {
 
 struct IXamlRootView;
+class ExpressionAnimationStore;
 
 typedef unsigned int LiveReloadCallbackCookie;
 typedef unsigned int ErrorCallbackCookie;
@@ -28,6 +30,8 @@ struct ReactInstanceSettings
   bool UseLiveReload { false };
   bool UseDirectDebugger{ false };
   bool UseJsi { true };
+  bool EnableJITCompilation { true };
+  std::string DebugHost;
   std::string DebugBundlePath;
   facebook::react::NativeLoggingHook LoggingCallback;
   std::function<void(facebook::react::JSExceptionInfo&&)> JsExceptionCallback;
@@ -71,8 +75,13 @@ struct IReactInstance
   virtual const std::string& LastErrorMessage() const noexcept = 0;
 
   virtual void loadBundle(std::string&& jsBundleRelativePath) = 0;
-};
 
-void UpdateDevSettings(bool useWebDebugger, bool useLiveReload, bool reuseReactInstances);
+  // Test Hooks
+  virtual void SetXamlViewCreatedTestHook(std::function<void(react::uwp::XamlView)> testHook) = 0;
+  virtual void CallXamlViewCreatedTestHook(react::uwp::XamlView view) = 0;
+
+  virtual ExpressionAnimationStore& GetExpressionAnimationStore() = 0;
+
+};
 
 } }
