@@ -4,6 +4,7 @@
 #pragma once
 
 #include "ViewManagerBase.h"
+#include <winrt/Windows.UI.Composition.h>
 
 namespace react { namespace uwp {
 
@@ -16,10 +17,17 @@ public:
   folly::dynamic GetNativeProps() const override;
   void UpdateProperties(ShadowNodeBase* nodeToUpdate, const folly::dynamic& reactDiffMap) override;
 
+  // Helper functions related to setting/updating TransformMatrix
+  void RefreshTransformMatrix(ShadowNodeBase* shadowNode);
+  void StartTransformAnimation(winrt::UIElement uielement, winrt::Windows::UI::Composition::CompositionPropertySet transformPS);
+
 protected:
   virtual void TransferProperties(XamlView oldView, XamlView newView) override;
-
   void TransferProperty(XamlView oldView, XamlView newView, winrt::Windows::UI::Xaml::DependencyProperty dp);
-};
+
+  void AnnounceIfNeeded(winrt::FrameworkElement element);
+
+private:
+  void ApplyTransformMatrix(winrt::UIElement uielement, ShadowNodeBase* shadowNode, winrt::Windows::Foundation::Numerics::float4x4 transformMatrix);};
 
 } }
